+++ lib/pragma_design_system.dart
@@ -2,9 +2,12 @@
 /// del Design System de Pragma.
 library pragma_design_system;
 
+export 'src/layout/pragma_grid.dart';
+export 'src/layout/pragma_grid_container.dart';
 export 'src/theme/pragma_theme.dart';
 export 'src/tokens/pragma_color_tokens.dart';
 export 'src/tokens/pragma_colors.dart';
+export 'src/tokens/pragma_grid_tokens.dart';
 export 'src/tokens/pragma_spacing.dart';
 export 'src/tokens/pragma_typography.dart';
 export 'src/widgets/pragma_button.dart';